rename template_nlg to template
