+++ tatk/nlg/template/camrest/__init__.py
@@ -1,0 +1,1 @@
+from tatk.nlg.template.camrest.nlg import TemplateNLG
