+++ tatk/nlg/template/camrest/auto_system_template_nlg.json
