+++ tatk/nlg/template/camrest/manual_user_template_nlg.json
