+++ tatk/nlg/template/multiwoz/__init__.py
@@ -1,0 +1,1 @@
+from tatk.nlg.template.multiwoz.nlg import TemplateNLG
