--- tatk/nlg/template_nlg/camrest/__init__.py
@@ -1,1 +1,0 @@
-from tatk.nlg.template_nlg.camrest.nlg import TemplateNLG
