--- tatk/nlg/template_nlg/multiwoz/__init__.py
@@ -1,1 +1,0 @@
-from tatk.nlg.template_nlg.multiwoz.nlg import TemplateNLG
